apps/backend/requirements.txt
alembic==1.13.1
argon2-cffi==23.1.0
celery[redis]==5.3.6
fastapi==0.110.1
google-auth==2.33.0
google-auth-httplib2==0.2.0
httpx==0.27.0
jinja2==3.1.4
passlib[argon2]==1.7.4
psycopg[binary]==3.1.18
pyjwt[crypto]==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
redis==5.0.4
sqlalchemy[asyncio]==2.0.30
sqlmodel==0.0.16
tenacity==8.3.0
uvicorn[standard]==0.29.0
python-slugify==8.0.4
pendulum==3.0.0
